requirements.txt
minimalmodbus==2.1.1
pyserial==3.5
PySide6==6.8.0.1
PySide6_Addons==6.8.0.1
PySide6_Essentials==6.8.0.1
PyYAML==6.0.2
shiboken6==6.8.0.1
